% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NO}
\alias{NO}
\title{Noah's preferred options}
\usage{
NO()
}
\description{
Set options, system variables, etc. as I like them
}
\examples{
Sys.getenv("TZ")
getOption("stringsAsFactors")

NO()

Sys.getenv("TZ")
getOption("stringsAsFactors")
}

