% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightAlignOnInterval.R
\name{rightAlignOnInterval}
\alias{rightAlignOnInterval}
\title{Right align on interval}
\usage{
rightAlignOnInterval(timestamps, alignment)
}
\arguments{
\item{timestamps}{POSIX time stamps}

\item{alignment}{regular alignment in seconds}
}
\description{
Right-align on a regular interval. Include the aligned interval in the alignment.
}

